feat: Update Todolist file

+++ Day02-Calculator.sol
@@ -1,34 +1,32 @@
-// SPDX-License-Identifier: MIT 
+// SPDX-License-Identifier: MIT
 pragma solidity ^0.8.13;
 
 /**
- * require statement 
+ * require statement
  * math operations
  * functions
  */
 contract Calulator {
- uint result;
+    int256 result;
 
- function add(uint a, uint b) public {
-     result = a + b;
- }
+    function add(int256 a, int256 b) public {
+        result = a + b;
+    }
 
- function sub(uint a, uint b) public {
-     result = a - b;
- }
+    function sub(int256 a, int256 b) public {
+        result = a - b;
+    }
 
- function mul(uint a, uint b) public {
-     result = a * b;
- }
+    function mul(int256 a, int256 b) public {
+        result = a * b;
+    }
 
- function div(uint a, uint b) public {
-     require(b > 0, "The second parameter should be larger than 0");
-     result = a / b;
- }
+    function div(int256 a, int256 b) public {
+        require(b > 0, "The second parameter should be larger than 0");
+        result = a / b;
+    }
 
- function getResult() public view returns(uint){
-     return result;
- }
+    function getResult() public view returns (int256) {
+        return result;
+    }
 }
-
-
